XYPlot: cell display etc.
Former-commit-id: efe7b86cd8a0954ae17063af14f68efe8204153e
Former-commit-id: 15ad3237b8fc2beeb83d8aa3c6bf27377ee29c44

+++ java/src/edu/caltech/ipac/firefly/visualize/graph/XYPlotData.java
@@ -261,7 +261,7 @@ public class XYPlotData {
             }
         });
         if (meta.getXSize()>0&&meta.getYSize()>0) {
-            sampler.setXYRatio(meta.getXSize()/meta.getYSize());
+            sampler.setXYRatio(((float)meta.getXSize())/((float)meta.getYSize()));
             int maxPoints = (int)(meta.getXSize()*meta.getYSize()/25.0); // assuming 5 px symbol
             if (maxPoints < 4) maxPoints = 4;
             if (maxPoints > 6400) maxPoints = 6400;
